Rename presenter and concern, and make other changes requested in code review
Fix code climate warning

Fix code climate

Tweak method name
