+++ spec/services/stats/management_information_generator_spec.rb
@@ -4,6 +4,27 @@ RSpec.describe Stats::ManagementInformationGenerator do
   subject(:result) { described_class.call }
 
   let(:frozen_time) { Time.new(2015, 3, 10, 11, 44, 55) }
+  let(:report_columns) { ['Scheme',
+                          'Case number',
+                          'Supplier number',
+                          'Organisation',
+                          'Case type name',
+                          'Bill type',
+                          'Claim total',
+                          'Submission type',
+                          'Submitted at',
+                          'Allocated at',
+                          'Completed at',
+                          'Current or end state',
+                          'State reason code',
+                          'Rejection reason',
+                          'Case worker',
+                          'Disk evidence case',
+                          'Main defendant',
+                          'Maat reference',
+                          'Rep order issued date',
+                          'AF1/LF1 processed by',
+                          'Misc fees'] }
 
   context 'data generation' do
     subject(:contents) { result.content.split("\n")}
@@ -22,8 +43,8 @@ RSpec.describe Stats::ManagementInformationGenerator do
       expect(contents.size).to eq(valid_claims.size + 1)
     end
     
-    it 'has 21 columns' do
-      expect(contents.first.split(',').count).to eql 21
+    it 'has expected columns' do
+      expect(contents.first.split(',')).to match_array(report_columns)
     end
   end
